requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.4
Pillow==9.3.0
python-decouple==3.6
qrcode==7.3.1
sqlparse==0.4.3
